backend/pyproject.toml
[tool.poetry]
name = "comercio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.7"
djangorestframework = "^3.11.0"
psycopg2 = "^2.8.5"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
ipython = "^7.15.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
